+++ djangocms_content_expiry/filters.py
@@ -190,3 +190,43 @@ class ContentExpiryDateRangeFilter(DateRangeFilter):
             queryset = queryset.filter(expires__range=(default_gte, default_lte))
 
         return queryset
+
+
+class ComplianceNumberFilter(admin.SimpleListFilter):
+    template = 'djangocms_content_expiry/input_filter.html'
+    title = _("Compliance Number")
+    parameter_name = "compliance_number"
+
+    def lookups(self, request, model_admin):
+        """
+        Required to show the filter
+        :param request: Request object
+        :param model_admin: Content expiry admin object
+        :return: Empty tuple
+        """
+        return (),
+
+    def choices(self, changelist):
+        """
+        Since the filter accept input, we only need the "all" choice option
+        :param changelist: Changelist object
+        :return: Choice object containing display choices
+        """
+        choice = next(super().choices(changelist))
+        choice['query_parts'] = (
+            (k, v)
+            for k, v in changelist.get_filters_params().items()
+            if k != self.parameter_name
+        )
+        yield choice
+
+    def queryset(self, request, queryset):
+        """
+        Filtered queryset based on value provided
+        :param request: Request object
+        :param queryset: Content Expiry queryset
+        :return: Filtered queryset
+        """
+        if self.value():
+            return queryset.filter(compliance_number=self.value()).distinct()
+        return queryset